+++ src/Symfony/Component/PropertyInfo/Extractor/PhpStanExtractor.php
@@ -23,8 +23,12 @@ use PHPStan\PhpDocParser\Parser\TokenIterator;
 use PHPStan\PhpDocParser\Parser\TypeParser;
 use Symfony\Component\PropertyInfo\PhpStan\NameScopeFactory;
 use Symfony\Component\PropertyInfo\PropertyTypeExtractorInterface;
-use Symfony\Component\PropertyInfo\Type;
+use Symfony\Component\PropertyInfo\Type as LegacyType;
 use Symfony\Component\PropertyInfo\Util\PhpStanTypeHelper;
+use Symfony\Component\TypeInfo\Exception\UnsupportedException;
+use Symfony\Component\TypeInfo\Type;
+use Symfony\Component\TypeInfo\TypeContext\TypeContextFactory;
+use Symfony\Component\TypeInfo\TypeResolver\StringTypeResolver;
 
 /**
  * Extracts data using PHPStan parser.
@@ -40,6 +44,9 @@ final class PhpStanExtractor implements PropertyTypeExtractorInterface, Construc
     private PhpDocParser $phpDocParser;
     private Lexer $lexer;
     private NameScopeFactory $nameScopeFactory;
+
+    private StringTypeResolver $stringTypeResolver;
+    private TypeContextFactory $typeContextFactory;
 
     /** @var array<string, array{PhpDocNode|null, int|null, string|null, string|null}> */
     private array $docBlocks = [];
@@ -71,10 +78,17 @@ final class PhpStanExtractor implements PropertyTypeExtractorInterface, Construc
         $this->phpDocParser = new PhpDocParser(new TypeParser(new ConstExprParser()), new ConstExprParser());
         $this->lexer = new Lexer();
         $this->nameScopeFactory = new NameScopeFactory();
+        $this->stringTypeResolver = new StringTypeResolver();
+        $this->typeContextFactory = new TypeContextFactory($this->stringTypeResolver);
     }
 
+    /**
+     * @deprecated since Symfony 7.1, use "getType" instead
+     */
     public function getTypes(string $class, string $property, array $context = []): ?array
     {
+        trigger_deprecation('symfony/property-info', '7.1', 'The "%s()" method is deprecated, use "%s::getType()" instead.', __METHOD__, self::class);
+
         /** @var PhpDocNode|null $docNode */
         [$docNode, $source, $prefix, $declaringClass] = $this->getDocBlock($class, $property);
         $nameScope = $this->nameScopeFactory->create($class, $declaringClass);
@@ -129,7 +143,7 @@ final class PhpStanExtractor implements PropertyTypeExtractorInterface, Construc
                         continue 2;
                 }
 
-                $types[] = new Type(Type::BUILTIN_TYPE_OBJECT, $type->isNullable(), $resolvedClass, $type->isCollection(), $type->getCollectionKeyTypes(), $type->getCollectionValueTypes());
+                $types[] = new LegacyType(LegacyType::BUILTIN_TYPE_OBJECT, $type->isNullable(), $resolvedClass, $type->isCollection(), $type->getCollectionKeyTypes(), $type->getCollectionValueTypes());
             }
         }
 
@@ -141,11 +155,18 @@ final class PhpStanExtractor implements PropertyTypeExtractorInterface, Construc
             return $types;
         }
 
-        return [new Type(Type::BUILTIN_TYPE_ARRAY, false, null, true, new Type(Type::BUILTIN_TYPE_INT), $types[0])];
+        return [new LegacyType(LegacyType::BUILTIN_TYPE_ARRAY, false, null, true, new LegacyType(LegacyType::BUILTIN_TYPE_INT), $types[0])];
     }
 
+    /**
+     * @deprecated since Symfony 7.1, use "getTypeFromConstructor" instead
+     *
+     * @return LegacyType[]|null
+     */
     public function getTypesFromConstructor(string $class, string $property): ?array
     {
+        trigger_deprecation('symfony/property-info', '7.1', 'The "%s()" method is deprecated, use "%s::getTypeFromConstructor()" instead.', __METHOD__, self::class);
+
         if (null === $tagDocNode = $this->getDocBlockFromConstructor($class, $property)) {
             return null;
         }
@@ -160,6 +181,63 @@ final class PhpStanExtractor implements PropertyTypeExtractorInterface, Construc
         }
 
         return $types;
+    }
+
+    public function getType(string $class, string $property, array $context = []): ?Type
+    {
+        /** @var PhpDocNode|null $docNode */
+        [$docNode, $source, $prefix, $declaringClass] = $this->getDocBlock($class, $property);
+
+        if (null === $docNode) {
+            return null;
+        }
+
+        $typeContext = $this->typeContextFactory->createFromClassName($class, $declaringClass);
+
+        $tag = match ($source) {
+            self::PROPERTY => '@var',
+            self::ACCESSOR => '@return',
+            self::MUTATOR => '@param',
+            default => 'invalid',
+        };
+
+        $types = [];
+
+        foreach ($docNode->getTagsByName($tag) as $tagDocNode) {
+            if ($tagDocNode->value instanceof InvalidTagValueNode) {
+                continue;
+            }
+
+            if ($tagDocNode->value instanceof ParamTagValueNode && null === $prefix && $tagDocNode->value->parameterName !== '$'.$property) {
+                continue;
+            }
+
+            try {
+                $types[] = $this->stringTypeResolver->resolve((string) $tagDocNode->value->type, $typeContext);
+            } catch (UnsupportedException) {
+            }
+        }
+
+        if (!$type = $types[0] ?? null) {
+            return null;
+        }
+
+        if (!\in_array($prefix, $this->arrayMutatorPrefixes, true)) {
+            return $type;
+        }
+
+        return Type::list($type);
+    }
+
+    public function getTypeFromConstructor(string $class, string $property): ?Type
+    {
+        if (!$tagDocNode = $this->getDocBlockFromConstructor($class, $property)) {
+            return null;
+        }
+
+        $typeContext = $this->typeContextFactory->createFromClassName($class);
+
+        return $this->stringTypeResolver->resolve((string) $tagDocNode->type, $typeContext);
     }
 
     private function getDocBlockFromConstructor(string $class, string $property): ?ParamTagValueNode