+++ src/Symfony/Component/PropertyInfo/Extractor/ReflectionExtractor.php
@@ -19,9 +19,15 @@ use Symfony\Component\PropertyInfo\PropertyReadInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyTypeExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyWriteInfo;
 use Symfony\Component\PropertyInfo\PropertyWriteInfoExtractorInterface;
-use Symfony\Component\PropertyInfo\Type;
+use Symfony\Component\PropertyInfo\Type as LegacyType;
 use Symfony\Component\String\Inflector\EnglishInflector;
 use Symfony\Component\String\Inflector\InflectorInterface;
+use Symfony\Component\TypeInfo\Exception\UnsupportedException;
+use Symfony\Component\TypeInfo\Type;
+use Symfony\Component\TypeInfo\Type\CollectionType;
+use Symfony\Component\TypeInfo\TypeIdentifier;
+use Symfony\Component\TypeInfo\TypeResolver\TypeResolver;
+use Symfony\Component\TypeInfo\TypeResolver\TypeResolverInterface;
 
 /**
  * Extracts data using the reflection API.
@@ -61,9 +67,9 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
     public const ALLOW_MAGIC_CALL = 1 << 2;
 
     private const MAP_TYPES = [
-        'integer' => Type::BUILTIN_TYPE_INT,
-        'boolean' => Type::BUILTIN_TYPE_BOOL,
-        'double' => Type::BUILTIN_TYPE_FLOAT,
+        'integer' => TypeIdentifier::INT->value,
+        'boolean' => TypeIdentifier::BOOL->value,
+        'double' => TypeIdentifier::FLOAT->value,
     ];
 
     private array $mutatorPrefixes;
@@ -76,6 +82,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
     private InflectorInterface $inflector;
     private array $arrayMutatorPrefixesFirst;
     private array $arrayMutatorPrefixesLast;
+    private TypeResolverInterface $typeResolver;
 
     /**
      * @param string[]|null $mutatorPrefixes
@@ -92,6 +99,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         $this->propertyReflectionFlags = $this->getPropertyFlags($accessFlags);
         $this->magicMethodsFlags = $magicMethodsFlags;
         $this->inflector = $inflector ?? new EnglishInflector();
+        $this->typeResolver = TypeResolver::create();
 
         $this->arrayMutatorPrefixesFirst = array_merge($this->arrayMutatorPrefixes, array_diff($this->mutatorPrefixes, $this->arrayMutatorPrefixes));
         $this->arrayMutatorPrefixesLast = array_reverse($this->arrayMutatorPrefixesFirst);
@@ -132,8 +140,13 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         return $properties ? array_values($properties) : null;
     }
 
+    /**
+     * @deprecated since Symfony 7.1, use "getType" instead
+     */
     public function getTypes(string $class, string $property, array $context = []): ?array
     {
+        trigger_deprecation('symfony/property-info', '7.1', 'The "%s()" method is deprecated, use "%s::getType()" instead.', __METHOD__, self::class);
+
         if ($fromMutator = $this->extractFromMutator($class, $property)) {
             return $fromMutator;
         }
@@ -156,8 +169,15 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         return null;
     }
 
+    /**
+     * @deprecated since Symfony 7.1, use "getTypeFromConstructor" instead
+     *
+     * @return LegacyType[]|null
+     */
     public function getTypesFromConstructor(string $class, string $property): ?array
     {
+        trigger_deprecation('symfony/property-info', '7.1', 'The "%s()" method is deprecated, use "%s::getTypeFromConstructor()" instead.', __METHOD__, self::class);
+
         try {
             $reflection = new \ReflectionClass($class);
         } catch (\ReflectionException) {
@@ -177,6 +197,89 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         }
 
         return $types;
+    }
+
+    public function getType(string $class, string $property, array $context = []): ?Type
+    {
+        [$mutatorReflection, $prefix] = $this->getMutatorMethod($class, $property);
+
+        if ($mutatorReflection) {
+            try {
+                $type = $this->typeResolver->resolve($mutatorReflection->getParameters()[0]);
+
+                if (!$type instanceof CollectionType && \in_array($prefix, $this->arrayMutatorPrefixes, true)) {
+                    $type = Type::list($type);
+                }
+
+                return $type;
+            } catch (UnsupportedException) {
+            }
+        }
+
+        [$accessorReflection, $prefix] = $this->getAccessorMethod($class, $property);
+        if ($accessorReflection) {
+            try {
+                return $this->typeResolver->resolve($accessorReflection);
+            } catch (UnsupportedException) {
+            }
+        }
+
+        if ($context['enable_constructor_extraction'] ?? $this->enableConstructorExtraction) {
+            try {
+                $reflectionClass = new \ReflectionClass($class);
+                if ($type = $this->extractTypeFromConstructor($reflectionClass, $property)) {
+                    return $type;
+                }
+            } catch (\ReflectionException) {
+            }
+        }
+
+        try {
+            $reflectionClass = new \ReflectionClass($class);
+            $reflectionProperty = $reflectionClass->getProperty($property);
+        } catch (\ReflectionException) {
+            return null;
+        }
+
+        try {
+            return $this->typeResolver->resolve($reflectionProperty);
+        } catch (UnsupportedException) {
+        }
+
+        if (null === $defaultValue = ($reflectionClass->getDefaultProperties()[$property] ?? null)) {
+            return null;
+        }
+
+        $typeIdentifier = TypeIdentifier::from(static::MAP_TYPES[\gettype($defaultValue)] ?? \gettype($defaultValue));
+        $type = 'array' === $typeIdentifier->value ? Type::array() : Type::builtin($typeIdentifier);
+
+        if ($this->isNullableProperty($class, $property)) {
+            $type = Type::nullable($type);
+        }
+
+        return $type;
+    }
+
+    public function getTypeFromConstructor(string $class, string $property): ?Type
+    {
+        try {
+            $reflection = new \ReflectionClass($class);
+        } catch (\ReflectionException) {
+            return null;
+        }
+
+        if (!$reflectionConstructor = $reflection->getConstructor()) {
+            return null;
+        }
+        if (!$reflectionParameter = $this->getReflectionParameterFromConstructor($property, $reflectionConstructor)) {
+            return null;
+        }
+
+        try {
+            return $this->typeResolver->resolve($reflectionParameter);
+        } catch (UnsupportedException) {
+            return null;
+        }
     }
 
     private function getReflectionParameterFromConstructor(string $property, \ReflectionMethod $reflectionConstructor): ?\ReflectionParameter
@@ -405,7 +508,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
     }
 
     /**
-     * @return Type[]|null
+     * @return LegacyType[]|null
      */
     private function extractFromMutator(string $class, string $property): ?array
     {
@@ -423,7 +526,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         $type = $this->extractFromReflectionType($reflectionType, $reflectionMethod->getDeclaringClass());
 
         if (1 === \count($type) && \in_array($prefix, $this->arrayMutatorPrefixes, true)) {
-            $type = [new Type(Type::BUILTIN_TYPE_ARRAY, false, null, true, new Type(Type::BUILTIN_TYPE_INT), $type[0])];
+            $type = [new LegacyType(LegacyType::BUILTIN_TYPE_ARRAY, false, null, true, new LegacyType(LegacyType::BUILTIN_TYPE_INT), $type[0])];
         }
 
         return $type;
@@ -432,7 +535,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
     /**
      * Tries to extract type information from accessors.
      *
-     * @return Type[]|null
+     * @return LegacyType[]|null
      */
     private function extractFromAccessor(string $class, string $property): ?array
     {
@@ -446,7 +549,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         }
 
         if (\in_array($prefix, ['is', 'can', 'has'])) {
-            return [new Type(Type::BUILTIN_TYPE_BOOL)];
+            return [new LegacyType(LegacyType::BUILTIN_TYPE_BOOL)];
         }
 
         return null;
@@ -455,7 +558,7 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
     /**
      * Tries to extract type information from constructor.
      *
-     * @return Type[]|null
+     * @return LegacyType[]|null
      */
     private function extractFromConstructor(string $class, string $property): ?array
     {
@@ -511,7 +614,31 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
         $type = \gettype($defaultValue);
         $type = static::MAP_TYPES[$type] ?? $type;
 
-        return [new Type($type, $this->isNullableProperty($class, $property), null, Type::BUILTIN_TYPE_ARRAY === $type)];
+        return [new LegacyType($type, $this->isNullableProperty($class, $property), null, LegacyType::BUILTIN_TYPE_ARRAY === $type)];
+    }
+
+    private function extractTypeFromConstructor(\ReflectionClass $reflectionClass, string $property): ?Type
+    {
+        if (!$constructor = $reflectionClass->getConstructor()) {
+            return null;
+        }
+
+        foreach ($constructor->getParameters() as $parameter) {
+            if ($property !== $parameter->name) {
+                continue;
+            }
+
+            try {
+                return $this->typeResolver->resolve($parameter);
+            } catch (UnsupportedException) {
+            }
+        }
+
+        if ($parentClass = $reflectionClass->getParentClass()) {
+            return $this->extractTypeFromConstructor($parentClass, $property);
+        }
+
+        return null;
     }
 
     private function extractFromReflectionType(\ReflectionType $reflectionType, \ReflectionClass $declaringClass): array
@@ -530,14 +657,14 @@ class ReflectionExtractor implements PropertyListExtractorInterface, PropertyTyp
                 continue;
             }
 
-            if (Type::BUILTIN_TYPE_ARRAY === $phpTypeOrClass) {
-                $types[] = new Type(Type::BUILTIN_TYPE_ARRAY, $nullable, null, true);
+            if (LegacyType::BUILTIN_TYPE_ARRAY === $phpTypeOrClass) {
+                $types[] = new LegacyType(LegacyType::BUILTIN_TYPE_ARRAY, $nullable, null, true);
             } elseif ('void' === $phpTypeOrClass) {
-                $types[] = new Type(Type::BUILTIN_TYPE_NULL, $nullable);
+                $types[] = new LegacyType(LegacyType::BUILTIN_TYPE_NULL, $nullable);
             } elseif ($type->isBuiltin()) {
-                $types[] = new Type($phpTypeOrClass, $nullable);
+                $types[] = new LegacyType($phpTypeOrClass, $nullable);
             } else {
-                $types[] = new Type(Type::BUILTIN_TYPE_OBJECT, $nullable, $this->resolveTypeName($phpTypeOrClass, $declaringClass));
+                $types[] = new LegacyType(LegacyType::BUILTIN_TYPE_OBJECT, $nullable, $this->resolveTypeName($phpTypeOrClass, $declaringClass));
             }
         }