+++ src/Symfony/Component/Serializer/Normalizer/BackedEnumNormalizer.php
@@ -11,9 +11,9 @@
 
 namespace Symfony\Component\Serializer\Normalizer;
 
-use Symfony\Component\PropertyInfo\Type;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
+use Symfony\Component\TypeInfo\Type;
 
 /**
  * Normalizes a {@see \BackedEnum} enumeration to a string or an integer.
@@ -70,7 +70,7 @@ final class BackedEnumNormalizer implements NormalizerInterface, DenormalizerInt
         }
 
         if (!\is_int($data) && !\is_string($data)) {
-            throw NotNormalizableValueException::createForUnexpectedDataType('The data is neither an integer nor a string, you should pass an integer or a string that can be parsed as an enumeration case of type '.$type.'.', $data, [Type::BUILTIN_TYPE_INT, Type::BUILTIN_TYPE_STRING], $context['deserialization_path'] ?? null, true);
+            throw NotNormalizableValueException::createForUnexpectedDataType('The data is neither an integer nor a string, you should pass an integer or a string that can be parsed as an enumeration case of type '.$type.'.', $data, [Type::int(), Type::string()], $context['deserialization_path'] ?? null, true);
         }
 
         try {